+++ src/getMessages.h
@@ -15,27 +15,27 @@
 
 Rcpp::DataFrame getMessagesTemplate(MessageType& msg,
                                     std::string filename, 
-                                    unsigned long long startMsgCount = 0,
-                                    unsigned long long endMsgCount = 0,
-                                    unsigned long long bufferSize = 1e8,
+                                    int64_t startMsgCount = 0,
+                                    int64_t endMsgCount = 0,
+                                    int64_t bufferSize = 1e8,
                                     bool quiet = false);
 
 Rcpp::DataFrame getOrders(std::string filename, 
-                          unsigned long long startMsgCount = 0,
-                          unsigned long long endMsgCount = 0,
-                          unsigned long long bufferSize = 1e8,
+                          int64_t startMsgCount = 0,
+                          int64_t endMsgCount = 0,
+                          int64_t bufferSize = 1e8,
                           bool quiet = false);
 
 Rcpp::DataFrame getTrades(std::string filename, 
-                          unsigned long long startMsgCount = 0,
-                          unsigned long long endMsgCount = 0,
-                          unsigned long long bufferSize = 1e8,
+                          int64_t startMsgCount = 0,
+                          int64_t endMsgCount = 0,
+                          int64_t bufferSize = 1e8,
                           bool quiet = false);
 
 Rcpp::DataFrame getModifications(std::string filename, 
-                                 unsigned long long startMsgCount = 0,
-                                 unsigned long long endMsgCount = 0,
-                                 unsigned long long bufferSize = 1e8,
+                                 int64_t startMsgCount = 0,
+                                 int64_t endMsgCount = 0,
+                                 int64_t bufferSize = 1e8,
                                  bool quiet = false);
 
 #endif //GETMESSAGES_H
